% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction_factor.R
\name{correction_factor}
\alias{correction_factor}
\title{Correction Factor}
\usage{
correction_factor(x, norm_value = 1.006)
}
\arguments{
\item{x}{a numeric vector of refractive index values}

\item{norm_value}{a numeric value specifying the target value to normalise to}
}
\value{
a numeric vector of correction factors
}
\description{
Calculate the normalisation correction factor
}
