% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacl_sg_standards.R
\docType{data}
\name{nacl_sg_standards}
\alias{nacl_sg_standards}
\title{Specific Gravity Values for NaCl Standards}
\format{
A \code{tbl_df}
\describe{
\item{sample_name}{sample name}
\item{specific_gravity}{specific gravity}
}
}
\usage{
nacl_sg_standards
}
\description{
Specific Gravity Values for NaCl Standards
}
\keyword{datasets}
