% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_volumes.R
\name{normalisation_volumes}
\alias{normalisation_volumes}
\title{Normalisation Volumes}
\usage{
normalisation_volumes(x, total_volume = 500)
}
\arguments{
\item{x}{a numeric vector of correction factors}

\item{total_volume}{a numeric value specifying the total required volume (urine + water)}
}
\value{
a \code{tbl_df} of required urine and water volumes
}
\description{
Calculate the required volumes based on normalisation correction factors
}
