% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_sheet.R
\name{pooling_sheet}
\alias{pooling_sheet}
\title{Urine Pooling Sheet}
\usage{
pooling_sheet(ri_data, norm_value, volume)
}
\arguments{
\item{ri_data}{a \code{tbl_df} of \strong{sample_name}, \strong{value} and \strong{group}}

\item{norm_value}{a numeric value to normalise to}

\item{volume}{a numeric value for the total volume per sample}
}
\value{
a \code{tbl_df} of pooling volumes for each \strong{sample_name}
}
\description{
Urine Pooling Sheet
}
