% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_adjustments.R
\name{sampel_adjustments}
\alias{sampel_adjustments}
\title{Urine Sample Adjustments}
\usage{
sampel_adjustments(ri_data, norm_value, volume)
}
\arguments{
\item{ri_data}{a \code{tbl_df} of \strong{sample_name}, \strong{value}}

\item{norm_value}{a numeric value to normalise to}

\item{volume}{a numeric value for the total volume per sample}
}
\value{
a \code{tbl_df} of pooling volumes for each \strong{sample_name}
}
\description{
Urine Sample Adjustments
}
