% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urea_plate_index.R
\docType{data}
\name{urea_plate_index}
\alias{urea_plate_index}
\title{96 Well Plate Map for Total Urea Assay}
\format{
A \code{tbl_df}
\describe{
\item{WELL}{well position}
\item{NAME}{sample name}
\item{CLASS}{sample class}
\item{CONC}{concentration}

}
}
\usage{
urea_plate_index
}
\description{
96 Well Plate Map for Total Urea Assay
}
\keyword{datasets}
