% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normVols.R
\name{normVols}
\alias{normVols}
\title{Create normalisation volumes}
\usage{
normVols(df, volume)
}
\arguments{
\item{df}{a \code{data.frame} with the following columns
\describe{
\item{id}{Unique sample identifier}
\item{value}{A value for urine conenctration (refractive index, osmolarity, etc..)}
\item{cls}{The class which sample will be grouped into for pooling}
}}
}
\value{
a \code{data.frame} with the following columns
\item{id}{Unique sample identifier}
\item{value}{A value for urine conenctration (refractive index, osmolarity, etc..)}
\item{cls}{The class which sample will be grouped into for pooling}
\item{UrineVol}{The volume of urine required}
\item{WaterVol}{The volume of water required}
}
\description{
Calculate the required Urine:Water volume for creating pools from multiple single urine samples, based
on a measure of urine concentration (ie, refractive index)
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

