% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refractive_index_prediction.R
\name{refractive_index_prediction}
\alias{refractive_index_prediction}
\title{Predict RI values from HPLC-RI}
\usage{
refractive_index_prediction(gaml_file)
}
\arguments{
\item{gaml_file}{a GAML file containing HPLC-RI data. See micturate::plate_maps('urine_ri') for more details}
}
\value{
a \code{tbl_df} of predicted refractive index values
}
\description{
Predict RI values from HPLC-RI
}
